Rewrite `tenzir-ctl export` into pipeline
